crdutils: Move test helpers to dedicated file and export them
This is done to improve reusability - the helpers can now be imported from
different packages.
